Split the is number function so that numeric tests are separated from testing unicode strings for numeric values
There may occasions where a seqlen value is represented as a unicode string.  In theses cases we can still handle them by doing a separate test for a numeric value on a converted unicode string.
This test is only performed if the regular is_number test fails.
We can now handle unicodes strings such as ⅕ to get a numeric representation of 0.2